Remove the `complete` handler, make it go through the consent page
