+++ crates/handlers/src/oauth2/mod.rs
@@ -23,7 +23,6 @@ use mas_storage::{Clock, RepositoryAccess};
 use thiserror::Error;
 
 pub mod authorization;
-pub mod consent;
 pub mod device;
 pub mod discovery;
 pub mod introspection;
